# app/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.4.0"
  title: URJC gestiongrupos Backend PWA
# during dev, should point to your local machine
host: localhost:10081
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/xml
paths:
...
  /private/votes: 
    x-swagger-router-controller: vote.controller
    post: 
      tags: 
        - "Votes"
      description: Create Vote
      operationId: createVote
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token
          required: true 
          type: string 
        - in: body 
          name: voteData
          description: Vote data
          schema: 
            $ref: "#/definitions/VoteData"
      responses: 
        "201": 
          description: Success 
          schema: 
            $ref: "#/definitions/VoteDataResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json:
              {              
                "id": "vote-XXXXXX",
	              "idParticipant": "88h",

	              "creationDate": "2019-05-28T00:30:00.000Z",
	              "idCall": "call-ucvb83kzsjw0pfsep",
	              "proposalsVote": [
		              {"idVoted": "001", "option": "2"}, {"idVoted": "002", "option": "1"}, {"idVoted": "003", "option": "3"}
	              ],
              }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Vote Controller]:createVote
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Vote Controller]:createVote
    get: 
      tags: 
        - "Votes" 
      description: Get votes
      operationId: getVotes
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string 
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/VotesDataListResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: 
              {              
                "id": "vote-XXXXXX",
	              "idParticipant": "88h",
	              "creationDate": "2019-05-28T00:30:00.000Z",
	              "idCall": "call-ucvb83kzsjw0pfsep",
	              "proposalsVote": [
		              {"idVoted": "001", "option": "2"}, {"idVoted": "002", "option": "1"}, {"idVoted": "003", "option": "3"}
	              ],
              }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Vote Controller]:getVotes
        "404":
          description: Votes not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: No votes found
              description: No votes found in [Vote Controller]:getVotes
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Vote Controller]:getVotes
  /private/votes/{idVote}: 
    x-swagger-router-controller: vote.controller
    get: 
      tags: 
        - "Votes" 
      description: Get vote by Id
      operationId: getVoteById
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string
        - in: path
          name: idVote
          description: idVote
          required: true 
          type: string 
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/VoteDataResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: 
              {              
                "id": "vote-XXXXXX",
	              "idParticipant": "88h",
	              "creationDate": "2019-05-28T00:30:00.000Z",
	              "idCall": "call-ucvb83kzsjw0pfsep",
	              "proposalsVote": [
		              {"idVoted": "001", "option": "2"}, {"idVoted": "002", "option": "1"}, {"idVoted": "003", "option": "3"}
	              ],
              }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Vote Controller]:getVoteById
        "404":
          description: Vote not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: Vote not found
              description: Vote not found in [Vote Controller]:getVoteById
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Vote Controller]:getVoteById
  /private/voteIdByParameters/{idCall}: 
    x-swagger-router-controller: vote.controller
    get: 
      tags: 
        - "Votes" 
      description: Get voteId by Parameters
      operationId: getVoteIdByParameters
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string
        - in: path
          name: idCall
          description: idCall
          required: true 
          type: string
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/VoteParametersResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: 
              {              
                "id": "vote-XXXXXX"
              }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Vote Controller]:VoteParameters
        "404":
          description: Vote not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: Vote not found
              description: Vote not found in [Vote Controller]:VoteParameters
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Vote Controller]:VoteParameters
  /private/votesHistory/{idVote}: 
    x-swagger-router-controller: vote.controller
    get: 
      tags: 
        - "Votes" 
      description: Get vote by Id with history
      operationId: getVoteByIdWithHistory
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string
        - in: path
          name: idVote
          description: idVote
          required: true 
          type: string 
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/VoteDataHistoryResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: 
              {              
                "id": "vote-XXXXXX",
	              "idParticipant": "88h",
	              "creationDate": "2019-05-28T00:30:00.000Z",
	              "idCall": "call-ucvb83kzsjw0pfsep",
	              "proposalsVote": [
		              {"idVoted": "001", "option": "2"}, {"idVoted": "002", "option": "1"}, {"idVoted": "003", "option": "3"}
                ],
                "tx_id": "0eb8c8dd8df3c1b5fe0e82ec94ea5b9d0cbab9229435b10e60e17ad70e47e73f",
              }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Vote Controller]:getVoteByIdWithHistory
        "404":
          description: Vote not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: Vote not found
              description: Vote not found in [Vote Controller]:getVoteByIdWithHistory
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Vote Controller]:getVoteByIdWithHistory


  # ##################################################################
  # Calls
  # ##################################################################
...
definitions:
...
  VoteData: 
    type: object
    properties:
      idParticipant: 
        type: string 
      creationDate:
        type: string
        format: date-time
      idCall:
        type: string
      proposalsVote:
        type: array
        items:
          $ref: "#/definitions/IdvotedOptionObject"

  # ##################################################################
  # RESPONSES
  # ##################################################################

  # Error Responses
  InnerErrorResponse:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  ErrorResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"
        
  # Authentication Responses
...
  UnauthenticatedResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"

  # Authorization Responses
  UnauthorizedResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"

  # Vote Responses
  IdvotedOptionObject:
    type: object
    properties:
      idVoted:
        type: string
      option:
        type: string
  VoteDataResponse: 
    type: object      
    properties:
      id:
        type: string
      idParticipant: 
        type: string 
      creationDate:
        type: string
        format: date-time
      idCall:
        type: string
      proposalsVote:
        type: array
        items:
          $ref: "#/definitions/IdvotedOptionObject"
  VoteDataHistoryResponse: 
    type: object      
    properties:
      id:
        type: string
      idParticipant: 
        type: string 
      creationDate:
        type: string
        format: date-time
      idCall:
        type: string
      proposalsVote:
        type: array
        items:
          $ref: "#/definitions/IdvotedOptionObject"
      tx_id:
        type: string
  VotesDataListResponse: 
    type: object
    properties:
      totalElements:
        type: number
      elements:
        type: array
        items: 
          $ref: "#/definitions/VoteDataResponse"
  VoteParametersResponse:
    type: object
    properties:
      id:
        type: string

  # Calls Responses
...